Separate CSV download testing base into helper class
